+++ app/src/main/res/layout/activity_main.xml
@@ -27,11 +27,6 @@
             android:layout_height="wrap_content"
             android:background="?attr/colorPrimary" />
 
-        <Button
-            android:id="@+id/button_testAPI"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="recipe search" />
     </com.google.android.material.appbar.AppBarLayout>
 
     <androidx.viewpager.widget.ViewPager
